♻️ remove any types in code
