+++ frontend/src/components/PayoutTable/index.tsx
@@ -1,4 +1,4 @@
-import { Currency, Payment, PaymentStatus, Sortable } from "src/types";
+import { Payment, Sortable } from "src/types";
 import Table from "../Table";
 import usePaymentSorting, { SortingFields } from "src/hooks/usePaymentSorting";
 import PaymentLine from "./Line";
@@ -33,33 +33,6 @@ const PayoutTable: React.FC<PropsType> = ({ payments, payoutInfoMissing }) => {
       ))}
     </Table>
   );
-};
-
-// TODO: replace this any with GraphQL-generated ts types
-export const mapApiPaymentsToProps = (apiPayment: any): Payment => {
-  const amount = { value: apiPayment.amountInUsd, currency: Currency.USD };
-  const project = apiPayment.budget.project;
-  const reason = apiPayment.reason?.work_items?.at(0);
-  const requestedAt = apiPayment.requestedAt;
-  const getPaidAmount = (payments: { amount: number }[]) =>
-    payments.reduce((total: number, payment: { amount: number }) => total + payment.amount, 0);
-
-  return {
-    id: apiPayment.id,
-    requestedAt: requestedAt,
-    amount,
-    reason,
-    project: {
-      id: project.id,
-      title: project.projectDetails.name,
-      shortDescription: project.projectDetails.shortDescription,
-      logoUrl: project.projectDetails.logoUrl,
-    },
-    status:
-      getPaidAmount(apiPayment.payments) === apiPayment.amountInUsd
-        ? PaymentStatus.ACCEPTED
-        : PaymentStatus.WAITING_PAYMENT,
-  };
 };
 
 export default PayoutTable;
